code : изменен тип на дабл

+++ project/project.cpp
@@ -6,9 +6,9 @@
 int main()
 {
 	cout << "Enter A and B: ";
-	int a, b;
+	double a, b;
 	cin >> a >> b;
-	int max = a;
+	double max = a;
 	if (b >= a) {
 		max = b;
 	}
